Access implementation (Policy)

+++ app/Http/Controllers/QuestionsController.php
@@ -9,6 +9,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show'] ]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -78,13 +82,18 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        
+            $this->authorize('update', $question);
         //return page and passed it in the question object
    //     dd($question);
     //    if (\Gate::allows('update-question', $question)) {
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, 'Access denied');
         }
-        // return view('questions.edit', compact('question'));
+        */
+        
+         return view('questions.edit', compact('question'));
         
     }
 
@@ -97,9 +106,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize('update', $question);
+        /*
         if (\Gate::denies('update-question', $question)) {
             abort(403, 'Access denied');
-        }
+        }*/
         
         $question->update($request->only('title', 'body'));
         
@@ -113,10 +124,13 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {/*
         if (\Gate::denies('delete-question', $question)) {
             abort(403, 'Access denied');
         }
+       */
+        
+        $this->authorize('delete', $question);
         
         $question->delete();
